+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -497,6 +497,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed a bug where instanced shadergraph shaders wouldn't compile on PS4.
 - Fixed an issue related to the envlightdatasrt not being bound in recursive rendering.
 - Fixed shadow cascade tooltip when using the metric mode (case 1229232)
+- Fixed how the area light influence volume is computed to match rasterization.
 - Focus on Decal uses the extends of the projectors
 
 ### Changed